n indicator

+++ n day 3/task 4.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include<conio.h>
 int main() {
-   int Simple_interest;
+   float Simple_interest;
    int p,r,n;
  
    printf("Enter The num :");
@@ -12,6 +12,6 @@ int main() {
    scanf("%d",&n);
 
    Simple_interest = (p*r*n);
-   printf("The area of the given rectangle is %d", Simple_interest/100);
+   printf("The area of the given rectangle is %.2f", Simple_interest/100);
    return 0;
 }
